+++ src/main/java/roomescape/reservation/ReservationTimeManager.java
@@ -1,0 +1,7 @@
+package roomescape.reservation;
+
+import java.time.LocalDateTime;
+
+public interface ReservationTimeManager {
+    LocalDateTime todayCurrentTime();
+}
