+++ src/main/java/roomescape/theme/ThemeService.java
@@ -4,7 +4,6 @@ import java.time.LocalDate;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import roomescape.common.TimeManager;
 import roomescape.exception.custom.reason.theme.ThemeNotFoundException;
 import roomescape.exception.custom.reason.theme.ThemeUsedException;
 import roomescape.reservation.repository.ReservationJpaRepository;
@@ -22,7 +21,7 @@ public class ThemeService {
 
     private final ThemeRepository themeRepository;
     private final ReservationJpaRepository reservationJpaRepository;
-    private final TimeManager timeManager;
+    private final ThemeTimeManager timeManager;
 
     public ThemeResponse create(
             final ThemeRequest request
