+++ src/main/java/roomescape/theme/ThemeTimeManager.java
@@ -1,0 +1,7 @@
+package roomescape.theme;
+
+import java.time.LocalDate;
+
+public interface ThemeTimeManager {
+    LocalDate today();
+}
