+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -25,8 +25,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
 import roomescape.payment.PaymentManager;
-import roomescape.common.TimeManager;
-import roomescape.payment.PaymentRequest;
+import roomescape.infrastructure.TimeManager;
+import roomescape.payment.dto.PaymentRequest;
 import roomescape.exception.custom.reason.payment.PaymentConfirmException;
 import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotExistsMemberException;
